feat: close #69034520 实现个人信息功能

+++ backend/base/src/models/person_info.rs
@@ -11,7 +11,7 @@ pub struct Model {
     pub name: String,
     pub identity_card: String,
     pub preferred_seat_location: String,
-    pub user_id: i64,
+    pub user_id: i32,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
